コメント機能追加(予定)、テーブルのリレーション修正、画像周り修正

+++ demo4/src/main/java/com/example/repository/BulletinBoardRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.BulletinBoard;
 
-public interface BulletinBoardRepository extends JpaRepository<BulletinBoard, Long>{
+public interface BulletinBoardRepository extends JpaRepository<BulletinBoard, Integer>{
 	public BulletinBoard findById(int id);
 	public void deleteById(int id);
 }
